+++ src/pages/Admin/AdminProductNew.tsx
@@ -1,0 +1,16 @@
+import ProductForm from '@/features/Admin/ProductForm';
+
+const AdminProductNewPage = () => {
+  return (
+    <>
+      <div className='flex justify-between items-center'>
+        <h2 className='font-semibold text-xl'>상품 등록</h2>
+      </div>
+      <div>
+        <ProductForm />
+      </div>
+    </>
+  );
+}
+
+export default AdminProductNewPage;
