+++ src/router/routes.tsx
@@ -17,7 +17,7 @@ import MypagePage from '../pages/User/Mypage';
 import MyinfoPage from '../pages/User/Myinfo';
 import AdminOrdersPage from '../pages/Admin/AdminOrders';
 import AdminProductsPage from '../pages/Admin/AdminProducts';
-import AdminProductsNewPage from '../pages/Admin/AdminProductsNew';
+import AdminProductNewPage from '../pages/Admin/AdminProductNew';
 
 const routes = createBrowserRouter([
   {
@@ -115,7 +115,7 @@ const routes = createBrowserRouter([
       },
       {
         path: 'product/new',
-        element: <AdminProductsNewPage />,
+        element: <AdminProductNewPage />,
       },
     ],
   },
